+++ termroom/__init__.py
@@ -1,3 +1,3 @@
 """Termroom implementation package."""
 
-__version__ = "0.3.9"
+__version__ = "0.3.10"